[refactor]: FileContainer 컴포넌트로 분리
댓글폼에서 재활용 가능하기 때문에 컴포넌트로 분리하였습니다.
